requirements.txt
nibabel>=4.0.0
nilearn>=0.10.0
antspyx>=0.3.8
numpy>=1.24.0
pandas>=2.0.0
matplotlib>=3.7.0
seaborn>=0.12.0
jupyter>=1.0.0
scipy>=1.10.0
